moved updateItem to backend
